refactor(backend): extract organization crud module
